server/server_django_ninja_version/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
arxiv==2.1.0
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
deep-translator==1.11.4
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
django-ninja==1.1.0
feedparser==6.0.10
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.12.0
pillow==10.3.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
sgmllib3k==1.0.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
